feat:实现jdk、json、kryo、hessian序列化器，动态使用

+++ example-common/src/main/java/com/qi/example/common/service/UserService.java
@@ -5,7 +5,7 @@ import com.qi.example.common.model.User;
 public interface UserService {
     User getUser(User user);
 
-    default short getNumber(){
+    default int getNumber(){
         return 1;
     }
 }
