+++ example-consumer/src/main/java/com/qi/example/consumer/ConsumerExample.java
@@ -19,7 +19,7 @@ public class ConsumerExample {
         } else {
             System.out.println("user == null");
         }
-        short number = userService.getNumber();
+        int number = userService.getNumber();
         System.out.println(number);
     }
 }
